[Refactor] Split column_viewer.h
