+++ be/src/exprs/vectorized/time_functions.cpp
@@ -5,6 +5,7 @@
 #include <string_view>
 
 #include "column/column_helper.h"
+#include "column/column_viewer.h"
 #include "exprs/vectorized/binary_function.h"
 #include "exprs/vectorized/unary_function.h"
 #include "runtime/datetime_value.h"
